+++ debug.log
@@ -1,0 +1,1 @@
+[1125/153423.500:ERROR:directory_reader_win.cc(43)] FindFirstFile: The system cannot find the path specified. (0x3)
