+++ src/components/RestaurantUI/RestaurantUI.css
@@ -1,6 +1,6 @@
 body {
   background-image:
-    url("../images/bg1.jpg");
+    url("../../images/bg1.jpg");
   background-size: cover;
 }
